requirements.txt
streamlit>=1.28.1
openai>=1.30.0
python-dotenv==1.0.0
pandas>=2.1.3
datetime
httpx
